Integrate boost solver. Later.
